backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# PDF processing
pdfminer.six>=20221105
reportlab>=4.0.0
pillow>=10.0.0

# Data processing & visualization
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
python-dateutil>=2.8.2
matplotlib>=3.8.0

# LLM and ML
groq>=0.4.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Azure Storage
azure-storage-blob>=12.18.0

# Salesforce
simple-salesforce>=1.12.9

# Google Cloud (optional, remove if not needed)
google-cloud-aiplatform>=1.38.0
google-cloud-logging>=3.7.0

# Performance
anyio>=3.7.1,<4.0.0
